update 3 tang

+++ DTO/TourDTO.cs
@@ -9,7 +9,7 @@ namespace DTO
     public class TourDTO
     {
         protected string _tourId, _tourPhoto, _tourName, _tourDescription, _tourDeparture, _tourDestination, _tourTime, _tourVehicle, _tourSchedule, _placeId;
-        protected int _tourPrice;
+        protected double _tourPrice;
         private string _status;
 
         public string TourId { get => _tourId; set => _tourId = value; }
@@ -23,13 +23,13 @@ namespace DTO
         public string TourSchedule { get => _tourSchedule; set => _tourSchedule = value; }
         public string PlaceId { get => _placeId; set => _placeId = value; }
         
-        public int TourPrice { get => _tourPrice; set => _tourPrice = value; }
+        public double TourPrice { get => _tourPrice; set => _tourPrice = value; }
         public string Status { get => _status; set => _status = value; }
         public TourDTO()
         {
 
         }
-        public TourDTO(string tourId, string tourPhoto, string tourName, string tourDescription, int tourPrice,
+        public TourDTO(string tourId, string tourPhoto, string tourName, string tourDescription, double tourPrice,
             string tourDeparture, string tourDestination, string tourTime, string tourVehicle, string tourSchedule,
             string placeId, string status)
         {
